impossible abilities set for each pkmn. start with adding intimidate on switch-in
Similar to impossible_items, this is a set containing abilities that
will not be guessed by the set inferral logic. We use this to make note
that certain abilities are impossible after certain events. In this
situation we will never guess intimidate after a pokemon has switched in
because intimidate would have revealed itself