More Zoroark edge-cases
If a pokemon is revealed to be zoroark then we need to reset the hp and
status values on the pokemon zoroark is disguised as. This is achieved
by setting hp and status values at switch-in based on what we _think_ the
pokemon should have based on the last known value of these attributes

This isn't perfect - this doesn't capture an edge-case where an opponent
pkmn switches into a Zoroark that is disguised as the pkmn switching
out. This should be strictly better than what existed before though