Fix compatibility with Pokemon Showdown
https://github.com/smogon/pokemon-showdown/commit/091241f1bbb2009b79f8e1a8e24f3b32d4e6150e
changed how |request| JSONs are sent and this project made a lot of
assumptions based on those JSONs coming first.

There are some situations where it is actually preferrable to parse the protocol _after_
the next-turn's request JSON is known so the implementation here stores
the protocol messages in a list until the request JSON is acquired and
only then are they parsed.

One such example:
```
|-formechange|p1a: Meloetta|Meloetta-Pirouette|300/350
|move|p2a: Ting-Lu|Earthquake|p1a: Meloetta
|-damage|p1a: Meloetta|44/350

|request|<json>  <-- contains forme-changed Meloetta-Priouette's stats
```

I want to know the next-turn's final stats of the forme-changed Meloetta-Pirouette
so I can do a reverse damage calc and figure out the EV ranges of Ting-Lu. This is only
possible with the request JSON for the next turn.

closes #250